+++ src/main/java/fuzs/betteranimationscollection/client/element/SoundModelElement.java
@@ -1,21 +1,18 @@
 package fuzs.betteranimationscollection.client.element;
 
-import com.google.common.collect.Lists;
 import fuzs.puzzleslib.config.ConfigManager;
 import fuzs.puzzleslib.config.option.OptionsBuilder;
 import fuzs.puzzleslib.config.serialization.EntryCollectionBuilder;
+import fuzs.puzzleslib.util.LoadedLocationList;
 import net.minecraft.entity.MobEntity;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundEvent;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 public abstract class SoundModelElement extends ModelElement {
 
-    protected final List<ResourceLocation> defaultSounds = Lists.newArrayList();
+    protected final LoadedLocationList defaultSounds = new LoadedLocationList();
 
     private Set<SoundEvent> sounds;
 
@@ -24,9 +21,7 @@ public abstract class SoundModelElement extends ModelElement {
     @Override
     public void setupModelConfig(OptionsBuilder builder) {
 
-        builder.define("Mob Sounds", this.defaultSounds.stream()
-                .map(ResourceLocation::toString)
-                .collect(Collectors.toList())).comment("Mob sounds to play a unique animation for.", "Useful for adding support for modded mob variants which have different sounds from their vanilla counterparts.", EntryCollectionBuilder.CONFIG_STRING).sync(v -> {
+        builder.define("Mob Sounds", this.defaultSounds).comment("Mob sounds to play a unique animation for.", "Useful for adding support for modded mob variants which have different sounds from their vanilla counterparts.", EntryCollectionBuilder.CONFIG_STRING).sync(v -> {
 
             if (this.sounds != null) {