+++ src/main/java/fuzs/betteranimationscollection/client/renderer/entity/model/OinkyPigModel.java
@@ -1,5 +1,7 @@
 package fuzs.betteranimationscollection.client.renderer.entity.model;
 
+import fuzs.betteranimationscollection.BetterAnimationsCollection;
+import fuzs.betteranimationscollection.client.element.OinkyPigElement;
 import net.minecraft.client.renderer.entity.model.PigModel;
 import net.minecraft.client.renderer.model.ModelRenderer;
 import net.minecraft.entity.Entity;
@@ -12,6 +14,8 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 public class OinkyPigModel<T extends Entity> extends PigModel<T> {
 
     private final ModelRenderer snout;
+    private final ModelRenderer earRight;
+    private final ModelRenderer earLeft;
 
     public OinkyPigModel() {
 
@@ -25,13 +29,23 @@ public class OinkyPigModel<T extends Entity> extends PigModel<T> {
         // overwrite normal head as it already includes the snout
         this.head = new ModelRenderer(this, 0, 0);
         this.head.addBox(-4.0F, -4.0F, -8.0F, 8, 8, 8, scale);
-        this.head.setPos(0.0F, (float) (18 - 6), -6.0F);
+        this.head.setPos(0.0F, 12.0F, -6.0F);
 
         // make snout a separate part
         this.snout = new ModelRenderer(this, 16, 16);
         this.snout.addBox(-2.0F, -3.0F, -1.0F, 4, 3, 1, scale);
         this.snout.setPos(0.0F, 3.0F, -8.0F);
         this.head.addChild(this.snout);
+
+        this.earRight = new ModelRenderer(this, 16, 4);
+        this.earRight.addBox(0.0F, 0.0F, -2.0F, 1.0F, 5.0F, 4.0F, scale);
+        this.earRight.setPos(3.5F, -2.0F, -4.0F);
+        this.head.addChild(this.earRight);
+        this.earLeft = new ModelRenderer(this, 16, 4);
+        this.earLeft.addBox(0.0F, 0.0F, -2.0F, 1.0F, 5.0F, 4.0F, scale);
+        this.earLeft.setPos(-3.5F, -2.0F, -4.0F);
+        this.earLeft.yRot = (float) Math.PI;
+        this.head.addChild(this.earLeft);
     }
 
     @Override
@@ -50,6 +64,25 @@ public class OinkyPigModel<T extends Entity> extends PigModel<T> {
 
                 this.snout.y = 3.0F;
             }
+        }
+    }
+
+    @Override
+    public void setupAnim(T entityIn, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
+
+        super.setupAnim(entityIn, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch);
+
+        OinkyPigElement element = (OinkyPigElement) BetterAnimationsCollection.OINKY_PIG;
+        boolean floatyEars = element.floatyEars;
+        this.earRight.visible = floatyEars;
+        this.earLeft.visible = floatyEars;
+
+        if (floatyEars) {
+
+            float f1 = ageInTicks * 0.1F + limbSwing * 0.5F;
+            float f2 = 0.08F + limbSwingAmount * element.earAnimationSpeed * 0.04F;
+            this.earRight.zRot = (-(float) Math.PI / 6.0F) - MathHelper.cos(f1 * 1.2F) * f2;
+            this.earLeft.zRot = ((float) Math.PI / 6.0F) + MathHelper.cos(f1) * f2;
         }
     }