+++ src/main/java/fuzs/betteranimationscollection/client/renderer/entity/model/SquidTentaclesModel.java
@@ -1,9 +1,9 @@
 package fuzs.betteranimationscollection.client.renderer.entity.model;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Iterables;
 import fuzs.betteranimationscollection.BetterAnimationsCollection;
 import fuzs.betteranimationscollection.client.element.SquidTentaclesElement;
-import fuzs.betteranimationscollection.client.util.ModelUtil;
 import net.minecraft.client.renderer.entity.model.SquidModel;
 import net.minecraft.client.renderer.model.ModelRenderer;
 import net.minecraft.entity.Entity;
@@ -23,7 +23,7 @@ public class SquidTentaclesModel<T extends Entity> extends SquidModel<T> {
 
         ImmutableList.Builder<ModelRenderer> builder = ImmutableList.builder();
         // get head from vanilla model
-        builder.add(ModelUtil.getAtIndex(super.parts().iterator(), 0));
+        builder.add(Iterables.get(super.parts(), 0));
         for (int j = 0; j < this.tentaclesStart.length; j++) {
 
             this.tentaclesStart[j] = new ModelRenderer(this, 48, 0);